config: extract config from main to separate package
