+++ config/config.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 import (
 	"encoding/base64"
